// client/src/boards/european-projects/pages/evolution-pcri/charts/countries-ranking/types.ts
export interface EvolutionDataItem {
  call_year: string;
  framework: string;
  country_name_fr: string;
  country_code: string;
  stage: string;
  number_involved: number;
  coordination_number: number;
  funding: number;
  project_number: number;
  share_number_involved: number;
  share_coordination_number: number;
  share_funding: number;
  share_project_number: number;
}